pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "audits-punishment"
version = "0.1.0"
description = "Data + LLM pipeline for linking Brazilian audit findings to legal punishment outcomes"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Audits Punishment Team" }
]
dependencies = [
  "pandas>=2.2.0",
  "pyarrow>=15.0.0",
  "requests>=2.32.0",
  "beautifulsoup4>=4.12.0",
  "lxml>=5.2.0",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.1",
  "tqdm>=4.66.0",
  "tenacity>=8.2.0",
  "loguru>=0.7.2",
  "pdfminer.six>=20231228"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.5.0"
]

[tool.hatch.build.targets.wheel]
packages = ["src/audits_punishment"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = []
